adds json converstion to cli output
This commit adds a new feature to the cli_get module that will now try
to convert the output from the command to json.  If the conversion is
successful then the json data structure is returned in the json key.  If
the conversion is not successful, then none is returned in the json key.

Regardless of the conversion attempt, the output is always returned in
the output key.