prefer json data over stdout when parsing with cli
If the parser argument is provided to the cli module, the output from
the cli command is parsed by the parser.  The cli module will attempt to
convert any output from the cli command from json.  This change will
prefer the json_data over the output for the cli command if the
json_data value is not none.