Create non-voting tox-linters job
This create a non-voting job using tox to run our linters entry point.
That in turn runs flake8 to validate our source is linted properly.
Because this is a non-voting job, we don't actually fail as follow up
patches will clean up the source code. Once done, we can make this job
voting and gating.

Signed-off-by: Paul Belanger <pabelanger@redhat.com>