adds new lookup plugins
* config_template
The config_template lookup plugin is derived from the template plugin
that is designed to work with network device configs.  It will
automationcally enable J2 trim_blocks and lstrip_blocks.   It also
supports the omit var

* network_config
The network_config lookup plugin will load and generate a device
configuration based on a set of directives defined in the network_config
file.