adds new directive extend (#91)
* adds new directive extend

This change adds a new directive to command_parser directives to extend
a set of fact subkeys.  If the extended keys do not exist then they are
created.

* fix issues with merging facts correctly

This change fixes a number of bugs related to merging facts correctly
when the extend directive is provided.  The plugin will now correctly
handle fact merging

This change also includes test cases for testing the extension
functionality

* fix up sanity checking errors