fixes an issue with using the extend directive with a loop
This change addresses a problem that would effectively cause the
directive to ignore the extend keyword when the export value is a
looped.  In that case, the return fact would be registered to the top
level key not the extended one.  This change addresses that issue such
that extend is honored during fact expansion