adds new lookup plugin config_template
This change adds a new lookup plugin that is designed to simply writing
templates for line based files such as network configurations.  It adds
the capabilite to pass default(omit) on a line and if the requested
variable is not defined, the entire line (including the static text) is
removed.