adds new module valiate for handling input validate
This change adds a new module that is designed to handle role input
validation based on an Ansible module argument spec.  This module takes
a single input that provides the elements of the argument spec to
validate and then checks the current vars to make sure the argument spec
is valid.